pyproject.toml
[project]
name = "smart-cache-paralog-api"
version = "1.2.3"
authors = [{name = "Telicent Ltd", email = "opensource@telicent.io"}]
requires-python = "==3.12"
dependencies = [
    "fastapi-healthchecks==1.1.0",
    "fastapi[standard]==0.115.12",
    "toml==0.10.2",
    "pydantic==2.5.2",
    "python-dotenv==1.1.0",
    "SPARQLWrapper==2.0.0",
]
classifiers = [
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3 :: Only',
    'Topic :: Software Development',
    'Typing :: Typed',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: POSIX',
    'Operating System :: Unix',
    'Operating System :: MacOS',
]

[project.urls]
repository = "https://github.com/telicent-oss/smart-cache-paralog-api"

[project.optional-dependencies]
dev = [
    "pre-commit==3.6.2", 
    "ruff==0.11.4",
    "mypy==1.15.0",
    "cyclonedx-bom==4.4.3",
    "types-toml",
    "types-requests==2.31.0.20240311",
    # For dataloading
    "pandas==2.2.3",
    "python-geohash==0.8.5",
    "Geohash==1.0",
    "termcolor==3.0.1"
]

[tool.ruff]
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 120

[tool.mypy]
no_implicit_optional = false
ignore_missing_imports = true
packages = ["paralog"]

[tool.setuptools]
py-modules = []
